feat: audio import pipeline with preview, reimport, and Show-in-Explorer
- Add convertAudioToWav() using miniaudio with libvorbis/libopus custom
  backends to transcode any supported audio format to 16-bit PCM WAV
- Compile miniaudio decoder backends directly into the editor binary so
  decoding symbols are always available regardless of SDK build state
- Link audio codec libs (ogg, vorbis, vorbisfile, opus, opusfile) for
  MSVC and MinGW; remove /NODEFAULTLIB:msvcrtd since all deps now use /MD
- Add audio asset preview panel in inspector with Play/Stop toggle
- Add startAudioPreviewByUuid() for direct asset playback; resolve
  Library/ImportedAssets path before falling back to Assets/
- Set default listener on preview manager so non-spatialized playback works
- Add "Reimport" and "Show in Explorer" context menu items in project panel
- Fix signed/unsigned mismatch: selectedIndices changed to size_t

+++ include/panel_console.h
@@ -52,7 +52,7 @@ class PanelConsole
 		char inputBuf[256] = "";                      ///< Backing buffer for the command input field.
 		bool scrollToBottom = false;                  ///< When true, the view scrolls to the latest entry on next draw.
 
-		std::unordered_set<int> selectedIndices;     ///< Indices of all currently selected lines.
+		std::unordered_set<size_t> selectedIndices;  ///< Indices of all currently selected lines.
         size_t lastClickedIndex = -1;               ///< Index of the last clicked line, used as the anchor for shift-range selection.
 
         bool showCopiedTooltip = false;              ///< Whether the "copied to clipboard" tooltip is currently shown.
